extract scenes into their own files
